fix: resolve linting errors in Info.tsx and routes.tsx
